requirement.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
Deprecated==1.2.13
Django==4.0.2
gcloud==0.18.3
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==3.3
jwcrypto==1.0
oauth2client==4.1.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pyparsing==3.0.7
Pyrebase4==4.5.0
python-jwt==3.3.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.8
wrapt==1.13.3
